Journal_OwlNote/src/test/java/com/example/Journal_OwlNote/Controller/QuestionControllerTest.java
package com.example.Journal_OwlNote.Controller;

import com.example.Journal_OwlNote.Model.JournalResponse;
import com.example.Journal_OwlNote.Model.Question;
import com.example.Journal_OwlNote.repo.QuestionRepository;
import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.data.ldap.DataLdapTest;
import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.ResultMatcher;
import static org.mockito.BDDMockito.given;
import java.util.Arrays;
import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.internal.bytebuddy.matcher.ElementMatchers.is;
import static org.junit.Assert.*;
import static org.hamcrest.Matchers.*;
//import static org.hamcrest.Ma
import static org.springframework.test.web.client.match.MockRestRequestMatchers.jsonPath;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;


@RunWith(SpringRunner.class)
@WebMvcTest(QuestionController.class)
public class QuestionControllerTest {

    @Autowired
    private MockMvc mvc;

    @MockBean
    private QuestionRepository service;

    @Test
    public void questions() {
    }

    @Test
    public void getById() {
    }

    @Test
    public void postQuestion() throws Exception {
        Question question1 = new Question("how old are you");


        List<Question> allQuestions = Arrays.asList(question1);

       // given(service.getAllQuestions()).willReturn(allQuestions);

        // then
        mvc.perform(get("/api/employees")
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect((ResultMatcher) jsonPath("$[5].question", is(question1.getQuestion())));
    }
}